move legacy command fixtures
