+++ tests/_support/Commands/Legacy/Foobar.php
